Refactor helm releases check
