+++ modules/340-monitoring-kubernetes/hooks/helm.go
@@ -17,26 +17,43 @@ limitations under the License.
 package hooks
 
 import (
+	"bytes"
+	"compress/gzip"
 	"context"
+	"encoding/base64"
+	"encoding/json"
 	"fmt"
+	"io"
 	"log"
+	"strings"
+	"sync"
+	"time"
 
 	"github.com/Masterminds/semver/v3"
 	"github.com/flant/addon-operator/pkg/module_manager/go_hook"
 	"github.com/flant/addon-operator/pkg/module_manager/go_hook/metrics"
+	"github.com/flant/addon-operator/pkg/utils/logger"
 	"github.com/flant/addon-operator/sdk"
-	"github.com/flant/kube-client/manifest/releaseutil" // nolint: staticcheck
+	"github.com/flant/shell-operator/pkg/kube_events_manager/types"
+	"github.com/golang/protobuf/proto" // nolint: staticcheck
+	"helm.sh/helm/v3/pkg/releaseutil"
+	v1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"sigs.k8s.io/yaml"
 
+	"github.com/deckhouse/deckhouse/dhctl/pkg/config"
 	"github.com/deckhouse/deckhouse/go_lib/dependency"
-	helmreleases "github.com/deckhouse/deckhouse/modules/340-monitoring-kubernetes/hooks/internal"
+	"github.com/deckhouse/deckhouse/go_lib/dependency/k8s"
+	"github.com/deckhouse/deckhouse/go_lib/dependency/requirements"
 )
 
 // this hook checks helm releases (v2 and v3) and find deprecated apis
 // hook returns only metrics:
 //   `resource_versions_compatibility` for apis:
 //      1 - is deprecated
-//      2 - in unsupported
+//      2 - is unsupported in the next k8s version inside the delta interval
+//		3 - is unsupported in the ANY next k8s version
 // Also hook returns count on deployed releases `helm_releases_count`
 // Hook checks only releases with status: deployed
 
@@ -88,6 +105,11 @@ const (
 	delta = 2
 	// objectBatchSize - how many secrets to list from k8s at once
 	objectBatchSize = int64(10)
+	// fetchSecretsInterval pause between fetching the helm secrets from apiserver
+	// need for avoiding apiserver overload
+	fetchSecretsInterval = 3 * time.Second
+
+	K8sVersionsWithDeprecations = "monitoringKubernetes:k8sVersionsWithDeprecations"
 )
 
 var helmStorage unsupportedVersionsStore
@@ -99,32 +121,59 @@ func init() {
 	}
 }
 
-// maximum time deep for cached releases. Variable required for overriding in tests.
-var helmReleasesInterval = helmreleases.IntervalHours1
-
 var _ = sdk.RegisterFunc(&go_hook.HookConfig{
-	Queue: "/modules/monitoring-kubernetes/helm_releases",
+	Queue: "/modules/monitoring-kubernetes/helm-releases-scan",
 	Schedule: []go_hook.ScheduleConfig{
 		{
 			Name:    "helm_releases",
 			Crontab: "0 * * * *", // every hour
 		},
 	},
-	OnStartup: &go_hook.OrderedConfig{
-		Order: 1,
+	Kubernetes: []go_hook.KubernetesConfig{
+		{
+			Name:              "kubernetesVersion",
+			ApiVersion:        "v1",
+			Kind:              "Secret",
+			NamespaceSelector: &types.NamespaceSelector{NameSelector: &types.NameSelector{MatchNames: []string{"kube-system"}}},
+			NameSelector:      &types.NameSelector{MatchNames: []string{"d8-cluster-configuration"}},
+			FilterFunc:        applyClusterConfigurationYamlFilter,
+		},
 	},
-}, dependency.WithExternalDependencies(handleHelmReleasesBySchedule))
+	// we don't need the startup hook, because this hook will start on synchronization
+}, dependency.WithExternalDependencies(handleHelmReleases))
 
-func handleHelmReleasesBySchedule(input *go_hook.HookInput, dc dependency.Container) error {
-	return handleHelmReleases(input, dc, helmReleasesInterval)
+func applyClusterConfigurationYamlFilter(obj *unstructured.Unstructured) (go_hook.FilterResult, error) {
+	secret := &v1.Secret{}
+	err := sdk.FromUnstructured(obj, secret)
+	if err != nil {
+		return nil, err
+	}
+
+	ccYaml, ok := secret.Data["cluster-configuration.yaml"]
+	if !ok {
+		return nil, fmt.Errorf(`"cluster-configuration.yaml" not found in "d8-cluster-configuration" Secret`)
+	}
+
+	var metaConfig *config.MetaConfig
+	metaConfig, err = config.ParseConfigFromData(string(ccYaml))
+	if err != nil {
+		return nil, err
+	}
+
+	kubernetesVersion, err := rawMessageToString(metaConfig.ClusterConfig["kubernetesVersion"])
+	if err != nil {
+		return nil, err
+	}
+
+	return kubernetesVersion, err
+}
+func rawMessageToString(message json.RawMessage) (string, error) {
+	var result string
+	err := json.Unmarshal(message, &result)
+	return result, err
 }
 
-func handleHelmReleases(input *go_hook.HookInput, dc dependency.Container, interval helmreleases.Interval) error {
-	var (
-		totalHelm3Releases uint32
-		totalHelm2Releases uint32
-	)
-
+func handleHelmReleases(input *go_hook.HookInput, dc dependency.Container) error {
 	input.MetricsCollector.Expire("helm_deprecated_apiversions")
 
 	k8sCurrentVersionRaw, ok := input.Values.GetOk("global.discovery.kubernetesVersion")
@@ -134,76 +183,115 @@ func handleHelmReleases(input *go_hook.HookInput, dc dependency.Container, inter
 	}
 	k8sCurrentVersion := semver.MustParse(k8sCurrentVersionRaw.String())
 
-	// create buffered channel == objectBatchSize
-	// this give as ability to handle in memory only objectBatchSize * 2 amount of helm releases
-	// because this counter also used as a limit to apiserver
-	// we have `objectBatchSize` (25) objects in channel and max `objectBatchSize` (25) objects in goroutine waiting for channel
-	releasesC := make(chan *helmreleases.Release, objectBatchSize)
-	doneC := make(chan bool)
+	var isAutomaticK8s bool
+	kubernetesVersion, ok := input.Snapshots["kubernetesVersion"]
+	if ok && len(kubernetesVersion) > 0 && kubernetesVersion[0].(string) == "Automatic" {
+		isAutomaticK8s = true
+		requirements.SaveValue(K8sVersionsWithDeprecations, "initial")
+	}
 
-	// helm3 and helm2 are listed and parsed in goroutines
-	// deprecated resources will be processed here in the separated goroutine
-	go runReleaseProcessor(k8sCurrentVersion, input, releasesC, doneC)
-
-	ctx := context.Background()
 	client, err := dc.GetK8sClient()
 	if err != nil {
 		return err
 	}
 
-	go func() {
-		totalHelm3Releases, totalHelm2Releases, err = helmreleases.GetHelmReleases(ctx, client, releasesC, interval)
+	ctx, cancel := context.WithCancelCause(context.Background())
+
+	processor := &helmDeprecatedAPIsProcessor{
+		logger: input.LogEntry,
+		ctx:    ctx,
+		cancel: cancel,
+	}
+
+	manifestsC := processor.FetchHelmManifests(client)
+	defer func() {
+		input.MetricsCollector.Set("helm_releases_count", float64(processor.totalHelm3Releases), map[string]string{"helm_version": "3"})
+		input.MetricsCollector.Set("helm_releases_count", float64(processor.totalHelm2Releases), map[string]string{"helm_version": "2"})
 	}()
-	<-doneC
 
-	// to avoid data race
-	input.MetricsCollector.Set("helm_releases_count", float64(totalHelm3Releases), map[string]string{"helm_version": "3"})
-	input.MetricsCollector.Set("helm_releases_count", float64(totalHelm2Releases), map[string]string{"helm_version": "2"})
+	// helm3 and helm2 are listed and parsed in goroutines
+	// deprecated resources will be processed here
+	deprecations, err := runManifestsCheck(ctx, k8sCurrentVersion, input, manifestsC)
+	if err != nil {
+		return err
+	}
 
+	if isAutomaticK8s {
+		if deprecations != "" {
+			requirements.SaveValue(K8sVersionsWithDeprecations, deprecations)
+		} else {
+			requirements.RemoveValue(K8sVersionsWithDeprecations)
+		}
+
+		return nil
+	}
+
+	requirements.RemoveValue(K8sVersionsWithDeprecations)
 	return nil
 }
 
-func runReleaseProcessor(k8sCurrentVersion *semver.Version, input *go_hook.HookInput, releasesC <-chan *helmreleases.Release, doneC chan<- bool) {
-	defer func() {
-		doneC <- true
-	}()
-	for rel := range releasesC {
-		for _, manifestData := range releaseutil.SplitManifests(rel.Manifest) {
-			resource := new(manifest)
-			err := yaml.Unmarshal([]byte(manifestData), &resource)
-			if err != nil {
-				input.LogEntry.Errorf("manifest (%s/%s) read error: %s", rel.Namespace, rel.Name, err)
-				continue
-			}
+func runManifestsCheck(ctx context.Context, k8sCurrentVersion *semver.Version, input *go_hook.HookInput, manifestsC <-chan *manifestHead) (string, error) {
+	allK8sWithDeprecations := make(map[string]struct{})
 
-			if resource == nil {
-				continue
+loop:
+	for {
+		select {
+		case resource, ok := <-manifestsC:
+			if !ok {
+				break loop
 			}
 
 			incompatibility, k8sCompatibilityVersion := helmStorage.CalculateCompatibility(k8sCurrentVersion, resource.APIVersion, resource.Kind)
-			if incompatibility > 0 {
+			switch incompatibility {
+			case UpToDateVersion:
+				// pass
+
+			case UnsupportedVersion, DeprecatedVersion:
+				allK8sWithDeprecations[k8sCompatibilityVersion] = struct{}{}
 				input.MetricsCollector.Set("resource_versions_compatibility", float64(incompatibility), map[string]string{
-					"helm_release_name":      rel.Name,
-					"helm_release_namespace": rel.Namespace,
-					"helm_version":           rel.HelmVersion,
+					"helm_release_name":      resource.HelmReleaseInfo.Name,
+					"helm_release_namespace": resource.HelmReleaseInfo.Namespace,
+					"helm_version":           resource.HelmReleaseInfo.Version,
 					"k8s_version":            k8sCompatibilityVersion,
 					"resource_name":          resource.Metadata.Name,
 					"resource_namespace":     resource.Metadata.Namespace,
 					"kind":                   resource.Kind,
 					"api_version":            resource.APIVersion,
 				}, metrics.WithGroup("helm_deprecated_apiversions"))
+
+			case FutureDeprecatedVersion:
+				allK8sWithDeprecations[k8sCompatibilityVersion] = struct{}{}
 			}
+
+		case <-ctx.Done():
+			if ctx.Err() != nil {
+				return "", ctx.Err()
+			}
+			break loop
 		}
 	}
+
+	result := strings.Builder{}
+	for k := range allK8sWithDeprecations {
+		result.WriteString(k + ",")
+	}
+
+	return strings.TrimSuffix(result.String(), ","), nil
 }
 
-type manifest struct {
+type manifestHead struct {
 	APIVersion string `json:"apiVersion" yaml:"apiVersion"`
 	Kind       string `json:"kind" yaml:"kind"`
 	Metadata   struct {
 		Name      string `json:"name" yaml:"name"`
 		Namespace string `json:"namespace" yaml:"namespace"`
 	} `json:"metadata" yaml:"metadata"`
+
+	HelmReleaseInfo struct {
+		Name      string
+		Namespace string
+		Version   string
+	} `json:"-" yaml:"-"`
 }
 
 // k8s version: APIVersion: [Kind, ...]
@@ -217,16 +305,19 @@ func (uvs unsupportedVersionsStore) getByK8sVersion(version *semver.Version) (un
 
 const (
 	UpToDateVersion uint = iota
+	// DeprecatedVersion marks deprecated resources that will be removed in the next delta releases
 	DeprecatedVersion
 	UnsupportedVersion
+	// FutureDeprecatedVersion resource that is deprecated in any future version
+	FutureDeprecatedVersion
 )
 
 // CalculateCompatibility check compatibility. Returns
 //
-//	 0 - if resource is compatible
-//	 1 - if resource in deprecated and will be removed in the future
-//	 2 - if resource is unsupported for current k8s version
-//	and k8s version in which deprecation would be
+//	0 - if resource is compatible
+//	1 - if resource in deprecated in the next delta(2) kubernetes versions and will be removed in the future
+//	2 - if resource is unsupported for current k8s version
+//	3 - if resource is deprecated in ANY k8s versions
 func (uvs unsupportedVersionsStore) CalculateCompatibility(currentVersion *semver.Version, resourceAPIVersion, resourceKind string) (uint, string) {
 	// check unsupported api for the current k8s version
 	currentK8SAPIsStorage, exists := uvs.getByK8sVersion(currentVersion)
@@ -237,16 +328,27 @@ func (uvs unsupportedVersionsStore) CalculateCompatibility(currentVersion *semve
 		}
 	}
 
-	// if api is supported - check deprecation in the next 2 minor k8s versions
-	for i := 1; i <= delta; i++ {
-		newMinor := currentVersion.Minor() + uint64(i)
-		nextVersion := semver.MustParse(fmt.Sprintf("%d.%d.0", currentVersion.Major(), newMinor))
-		storage, exists := uvs.getByK8sVersion(nextVersion)
-		if exists {
-			isDeprecated := storage.isUnsupportedByAPIAndKind(resourceAPIVersion, resourceKind)
-			if isDeprecated {
-				return DeprecatedVersion, fmt.Sprintf("%d.%d", nextVersion.Major(), nextVersion.Minor())
-			}
+	for version, store := range helmStorage {
+		foundK8sVersion := semver.MustParse(version)
+		if currentVersion.GreaterThan(foundK8sVersion) {
+			// skip deprecation in previous k8s versions
+			continue
+		}
+		if !store.isUnsupportedByAPIAndKind(resourceAPIVersion, resourceKind) {
+			// skip up-to-date resources
+			continue
+		}
+
+		resultString := fmt.Sprintf("%d.%d", foundK8sVersion.Major(), foundK8sVersion.Minor())
+		switch {
+		case foundK8sVersion.Minor() == currentVersion.Minor():
+			return UnsupportedVersion, resultString
+
+		case foundK8sVersion.Minor() <= currentVersion.Minor()+delta:
+			return DeprecatedVersion, resultString
+
+		case foundK8sVersion.Minor() > currentVersion.Minor()+delta:
+			return FutureDeprecatedVersion, resultString
 		}
 	}
 
@@ -269,3 +371,258 @@ func (ua unsupportedAPIVersions) isUnsupportedByAPIAndKind(resourceAPI, resource
 
 	return false
 }
+
+// Block with helm release deprecated api processor
+type helmDeprecatedAPIsProcessor struct {
+	totalHelm3Releases uint32
+	totalHelm2Releases uint32
+	logger             logger.Logger
+
+	ctx    context.Context
+	cancel context.CancelCauseFunc
+}
+
+func (h *helmDeprecatedAPIsProcessor) getHelm3Releases(client k8s.Client, releasesC chan<- *Release) error {
+	var next string
+
+	for {
+		secretsList, err := client.CoreV1().Secrets("").List(h.ctx, metav1.ListOptions{
+			LabelSelector: "owner=helm,status=deployed",
+			Limit:         objectBatchSize,
+			Continue:      next,
+			// https://kubernetes.io/docs/reference/using-api/api-concepts/#semantics-for-get-and-list
+			// set explicit behavior:
+			//   Return data at any resource version. The newest available resource version is preferred, but strong consistency is not required; data at any resource version may be served.
+			ResourceVersion:      "0",
+			ResourceVersionMatch: metav1.ResourceVersionMatchNotOlderThan,
+		})
+		if err != nil {
+			return err
+		}
+		secretsList.GetRemainingItemCount()
+
+		for _, secret := range secretsList.Items {
+			releaseData := secret.Data["release"]
+			if len(releaseData) == 0 {
+				continue
+			}
+
+			release, err := helm3DecodeRelease(string(releaseData))
+			if err != nil {
+				return err
+			}
+			// release can contain wrong namespace (set by helm and werf) and confuse user with a wrong metric
+			// fetch namespace from secret is more reliable
+			release.Namespace = secret.Namespace
+			release.HelmVersion = "3"
+
+			releasesC <- release
+			h.totalHelm3Releases++
+		}
+
+		if secretsList.Continue == "" {
+			break
+		}
+
+		next = secretsList.Continue
+		time.Sleep(fetchSecretsInterval)
+	}
+
+	return nil
+}
+func (h *helmDeprecatedAPIsProcessor) getHelm2Releases(client k8s.Client, releasesC chan<- *Release) error {
+	var next string
+
+	for {
+		cmList, err := client.CoreV1().ConfigMaps("").List(h.ctx, metav1.ListOptions{
+			LabelSelector:        "OWNER=TILLER,STATUS=DEPLOYED",
+			Limit:                objectBatchSize,
+			Continue:             next,
+			ResourceVersion:      "0",
+			ResourceVersionMatch: metav1.ResourceVersionMatchNotOlderThan,
+		})
+		if err != nil {
+			return err
+		}
+
+		for _, secret := range cmList.Items {
+			releaseData := secret.Data["release"]
+			if len(releaseData) == 0 {
+				continue
+			}
+
+			release, err := helm2DecodeRelease(releaseData)
+			if err != nil {
+				return err
+			}
+			// release can contain wrong namespace (set by helm and werf) and confuse user with a wrong metric
+			// fetch namespace from secret is more reliable
+			release.Namespace = secret.Namespace
+			release.HelmVersion = "2"
+
+			releasesC <- release
+			h.totalHelm2Releases++
+		}
+
+		if cmList.Continue == "" {
+			break
+		}
+
+		next = cmList.Continue
+		time.Sleep(fetchSecretsInterval)
+	}
+
+	return nil
+}
+
+func (h *helmDeprecatedAPIsProcessor) getHelmReleases(client k8s.Client) chan *Release {
+	var (
+		wg        sync.WaitGroup
+		releasesC = make(chan *Release, objectBatchSize*2)
+	)
+
+	wg.Add(2)
+	go func() {
+		defer wg.Done()
+		err := h.getHelm3Releases(client, releasesC)
+		if err != nil {
+			h.cancel(err)
+			return
+		}
+	}()
+
+	go func() {
+		defer wg.Done()
+		err := h.getHelm2Releases(client, releasesC)
+		if err != nil {
+			h.cancel(err)
+			return
+		}
+	}()
+
+	go func() {
+		wg.Wait()
+		close(releasesC)
+	}()
+
+	return releasesC
+}
+
+func (h *helmDeprecatedAPIsProcessor) FetchHelmManifests(client k8s.Client) chan *manifestHead {
+	manifestsC := make(chan *manifestHead, objectBatchSize*2)
+	releasesC := h.getHelmReleases(client)
+
+	go func() {
+		for rel := range releasesC {
+			if h.ctx.Err() != nil {
+				// return on cancelled context
+				return
+			}
+			for _, manifestData := range releaseutil.SplitManifests(rel.Manifest) {
+				resource := new(manifestHead)
+				err := yaml.Unmarshal([]byte(manifestData), &resource)
+				if err != nil {
+					h.logger.Warnf("manifest (%s/%s) read error: %s", rel.Namespace, rel.Name, err)
+					continue
+				}
+
+				if resource == nil {
+					continue
+				}
+
+				resource.HelmReleaseInfo.Name = rel.Name
+				resource.HelmReleaseInfo.Namespace = rel.Namespace
+				resource.HelmReleaseInfo.Version = rel.HelmVersion
+
+				manifestsC <- resource
+			}
+		}
+
+		close(manifestsC)
+	}()
+
+	return manifestsC
+}
+
+// helm decoding
+var magicGzip = []byte{0x1f, 0x8b, 0x08}
+
+// Import this from helm3 lib - https://github.com/helm/helm/blob/49819b4ef782e80b0c7f78c30bd76b51ebb56dc8/pkg/storage/driver/util.go#L56
+// helm3DecodeRelease decodes the bytes of data into a release
+// type. Data must contain a base64 encoded gzipped string of a
+// valid release, otherwise an error is returned.
+func helm3DecodeRelease(data string) (*Release, error) {
+	// base64 decode string
+	b, err := base64.StdEncoding.DecodeString(data)
+	if err != nil {
+		return nil, err
+	}
+
+	// For backwards compatibility with releases that were stored before
+	// compression was introduced we skip decompression if the
+	// gzip magic header is not found
+	if bytes.Equal(b[0:3], magicGzip) {
+		r, err := gzip.NewReader(bytes.NewReader(b))
+		if err != nil {
+			return nil, err
+		}
+		defer r.Close()
+		b2, err := io.ReadAll(r)
+		if err != nil {
+			return nil, err
+		}
+		b = b2
+	}
+
+	var rls Release
+	// unmarshal release object bytes
+	if err := json.Unmarshal(b, &rls); err != nil {
+		return nil, err
+	}
+	return &rls, nil
+}
+
+// https://github.com/helm/helm/blob/47f0b88409e71fd9ca272abc7cd762a56a1c613e/pkg/storage/driver/util.go#L57
+func helm2DecodeRelease(data string) (*Release, error) {
+	// base64 decode string
+	b, err := base64.StdEncoding.DecodeString(data)
+	if err != nil {
+		return nil, err
+	}
+
+	// For backwards compatibility with releases that were stored before
+	// compression was introduced we skip decompression if the
+	// gzip magic header is not found
+	if bytes.Equal(b[0:3], magicGzip) {
+		r, err := gzip.NewReader(bytes.NewReader(b))
+		if err != nil {
+			return nil, err
+		}
+		b2, err := io.ReadAll(r)
+		if err != nil {
+			return nil, err
+		}
+		b = b2
+	}
+
+	var rls Release
+	// unmarshal protobuf bytes
+	if err := proto.Unmarshal(b, &rls); err != nil {
+		return nil, err
+	}
+	return &rls, nil
+}
+
+// protobuf methods for helm2
+type Release struct {
+	Name      string `json:"name,omitempty" protobuf:"bytes,1,opt,name=name,proto3"`
+	Namespace string `json:"namespace,omitempty" protobuf:"bytes,8,opt,name=namespace,proto3"`
+	Manifest  string `json:"manifest,omitempty" protobuf:"bytes,5,opt,name=manifest,proto3"`
+
+	// set helm version manually
+	HelmVersion string `json:"-"`
+}
+
+func (m *Release) Reset()         { *m = Release{} }
+func (m *Release) String() string { return proto.CompactTextString(m) }
+func (*Release) ProtoMessage()    {}
